Selection.testPoint() exception on some mouseovers
At present, an exception occurs when mouse moves over a rect (and mouseover is bound to a selection transform). In that case, values[i] is undefined so "values[i][0]" throws.

Disclaimer: In my personal use cases, this patch is satisfactory, but I have not tested this widely, and I do not know the codebase. This patch avoids the current exception but might not have correct semantics: perhaps the function should return false when members of the "values" array are undefined, at least in some cases. Alternatively, perhaps the root cause is that the mouseover should not be passing undefined data. In addition to this PR, I'll also submit an issue and let you decide which one is redundant.

Exception in Chrome is as follows: // + file paths
logger.js:3 ERROR TypeError: Cannot read property '0' of undefined
    at xI (selection.js:30) // vega-parser/src/parsers/expression/
    at Object.vlSelectionTest (selection.js:97) // vega-parser/src/parsers/expression/
    at Object.eval (eval at oE (expression.js:9), <anonymous>:3:126) // vega-runtime
    at Encode.js:70 // vega-encode
    at Array.forEach (<anonymous>)
    at oe (visitArray.js:9) // vega-util
    at qr.Qr.visit (Pulse.js:390) // vega-dataflow
    at Ty.Object.transform (Encode.js:69) // vega-encode
    at Ty.ho.evaluate (Transform.js:52) // vega-dataflow
    at Ty.ho.run (Transform.js:33)  // vega-dataflow

Repro case follows. I tested in the Vega-Lite editor in latest Chrome and Edge on Win10. Note that you can replace the rule mark with a text mark and get the same error. You can also omit the y encoding from the rule. And, you can change x to a quantitative scale.

{
  "datasets": {
    "days": [
      { "DayString": "2018-10-30T00:00:00.000Z", "Label": "Tues" },
      { "DayString": "2018-10-31T00:00:00.000Z", "Label": "Wed" },
      { "DayString": "2018-11-01T00:00:00.000Z", "Label": "Thurs" } 
    ],
    "blocks": [ { "BlockDay": "2018-10-30T00:00:00.000Z", "End": "2018-11-01T00:00:00.000Z", "Bottom": 1, "Top": 3 } ]
  },
  "$schema": "https://vega.github.io/schema/vega-lite/v3.0.0-rc8.json",
  "layer": [
    {
      "data": {"name": "days"},
      "layer": [
        {
          "selection": {
            "daypick": {
              "type": "single",
              "nearest": true,
              "on": "mouseover",
              "encodings": ["x"],
              "empty": "none"
            }
          },
          "mark": "text",
          "encoding": {
            "text": { "field": "Label"},
            "x": { 
              "field": "DayString",
              "type": "temporal",
              "scale": {
                "domain": ["2018-10-28T00:00:00.000Z", "2018-11-02T00:00:00.000Z"]
              },
              "axis": {"format": "%-m/%-e", "title": "Day"}
            },
            "y" : { "value": 4, "type": "quantitative"},
            "opacity": {
              "condition": {"selection": "daypick", "value": 1},
              "value": 0
            }
          }
        }
      ]
    },
    {
      "data": {"name": "blocks"},
      "mark": {"type": "rect"},
      "encoding": {
        "x": { "field": "BlockDay", "type": "temporal" },
        "y": { "field": "Bottom", "type": "quantitative", "scale": {"domain": [0, 4]} },
        "x2": { "field": "End", "type": "temporal" },
        "y2": { "field": "Top", "type": "quantitative" }
      }
    }
  ]
}